Additional autoinjection code
This is completing the same functionality that the Autoinjection code
added yesterday was attempting to complete; however, this has known
delay timings for mL amounts for the syringe we are using. This will
need work to made more abstract. We can decide on which code to use in
the future.